Split the alignment column and alignment "matrix" stuff out into a separate file
This was more involved than I would have thought, and required pulling
the boost graph types out into a separate header, for example.
